feat: 워크스페이스 소유자 회원 탈퇴 제한
- 워크스페이스 OWNER인 유저는 탈퇴 불가하도록 검증 추가
- WORKSPACE_OWNER_CANNOT_WITHDRAW 에러코드 추가
- WorkspaceMemberRepository에 existsByUser_IdAndRole 메서드 추가
- 소유자 탈퇴 제한 테스트 추가

+++ src/main/java/com/ujax/application/user/UserService.java
@@ -6,7 +6,10 @@ import org.springframework.transaction.annotation.Transactional;
 import com.ujax.application.user.dto.response.UserResponse;
 import com.ujax.domain.user.User;
 import com.ujax.domain.user.UserRepository;
+import com.ujax.domain.workspace.WorkspaceMemberRepository;
+import com.ujax.domain.workspace.WorkspaceMemberRole;
 import com.ujax.global.exception.ErrorCode;
+import com.ujax.global.exception.common.BusinessRuleViolationException;
 import com.ujax.global.exception.common.NotFoundException;
 import com.ujax.infrastructure.web.user.dto.request.UserUpdateRequest;
 
@@ -18,6 +21,7 @@ import lombok.RequiredArgsConstructor;
 public class UserService {
 
 	private final UserRepository userRepository;
+	private final WorkspaceMemberRepository workspaceMemberRepository;
 
 	public UserResponse getUser(Long userId) {
 		return UserResponse.from(findUserById(userId));
@@ -33,6 +37,9 @@ public class UserService {
 	@Transactional
 	public void deleteUser(Long userId) {
 		User user = findUserById(userId);
+		if (workspaceMemberRepository.existsByUser_IdAndRole(userId, WorkspaceMemberRole.OWNER)) {
+			throw new BusinessRuleViolationException(ErrorCode.WORKSPACE_OWNER_CANNOT_WITHDRAW);
+		}
 		userRepository.delete(user);
 	}
